Added main html file
Blog is coded entirely in html with css and js embedded. It would be a shame if someone hit the "<> Code" button (bright green, can't miss it) downloaded the html file, (download .zip) then opened in in any text editor (online or local) (vscode, replit, webmaker) so that they could edit it to fit their liking and maybe get a good grade on a school coding project. (pls dont steal)